# backend/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MediSync API",
    "description": "Hospital Management System API for managing doctors, patients, and appointments.",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/doctors/{doctorId}": {
      "get": {
        "tags": [
          "Doctors"
        ],
        "summary": "Get doctor details",
        "description": "Retrieves public profile information for a specific doctor.",
        "operationId": "getDoctorById",
        "parameters": [
          {
            "name": "doctorId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved doctor",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorDto"
                }
              }
            }
          },
          "404": {
            "description": "Doctor not found"
          }
        }
      },
      "put": {
        "tags": [
          "Doctors"
        ],
        "summary": "Update doctor details",
        "description": "Updates profile information (specialization, etc.). Requires ADMIN role.",
        "operationId": "updateDoctor",
        "parameters": [
          {
            "name": "doctorId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DoctorUpdateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Doctor updated successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Requires ADMIN role"
          },
          "404": {
            "description": "Doctor not found"
          }
        }
      }
    },
    "/api/doctors/{doctorId}/deactivate": {
      "put": {
        "tags": [
          "Doctors"
        ],
        "summary": "Deactivate doctor",
        "description": "Soft-deletes a doctor, preventing new bookings. Requires ADMIN role.",
        "operationId": "deactivateDoctor",
        "parameters": [
          {
            "name": "doctorId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Doctor successfully deactivated (No Content)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Requires ADMIN role"
          },
          "404": {
            "description": "Doctor not found"
          }
        }
      }
    },
    "/api/doctors/{doctorId}/activate": {
      "put": {
        "tags": [
          "Doctors"
        ],
        "summary": "Activate doctor",
        "description": "Restores a deactivated doctor account. Requires ADMIN role.",
        "operationId": "activate",
        "parameters": [
          {
            "name": "doctorId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Doctor successfully activated (No Content)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Requires ADMIN role"
          },
          "404": {
            "description": "Doctor not found"
          }
        }
      }
    },
...
    "/api/doctors": {
      "get": {
        "tags": [
          "Doctors"
        ],
        "summary": "List all doctors",
        "description": "Retrieves a list of doctors with optional filtering.",
        "operationId": "getDoctors",
        "parameters": [
          {
            "name": "departmentId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "deactivated",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved list",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DoctorDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Doctors"
        ],
        "summary": "Register a new doctor",
        "description": "Creates a new doctor profile and user account. Requires ADMIN role.",
        "operationId": "registerDoctor",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DoctorRegistrationDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Doctor registered successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input (e.g., email already exists)"
          },
          "403": {
            "description": "Forbidden - Requires ADMIN role"
          }
        }
      }
    },
...
    "/api/doctors/{doctorId}/appointments": {
      "get": {
        "tags": [
          "Doctors"
        ],
        "summary": "Get doctor's appointments",
        "description": "Retrieves all appointments for a specific doctor. Restricted to the Doctor themselves or Admins.",
        "operationId": "getAppointmentsByDoctorId",
        "parameters": [
          {
            "name": "doctorId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved appointments",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AppointmentDto"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Requires DOCTOR or ADMIN role"
          },
          "404": {
            "description": "Doctor not found"
          }
        }
      }
    },
    "/api/doctors/{doctorId}/appointments/slots": {
      "get": {
        "tags": [
          "Doctors"
        ],
        "summary": "Get available time slots",
        "description": "Returns a list of available appointment times for a doctor on a specific date.",
        "operationId": "getAvailableSlots",
        "parameters": [
          {
            "name": "doctorId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved slots",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid date format"
          },
          "404": {
            "description": "Doctor not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PatientDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date"
          },
          "gender": {
            "type": "string",
            "enum": [
              "MALE",
              "FEMALE",
              "OTHER"
            ]
          },
          "address": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "county": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "MedicalRecordDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "diagnosis": {
            "type": "string"
          },
          "treatmentPlan": {
            "type": "string"
          },
          "prescription": {
            "type": "string"
          },
          "appointmentId": {
            "type": "integer",
            "format": "int64"
          },
          "appointmentTime": {
            "type": "string",
            "format": "date-time"
          },
          "doctorId": {
            "type": "integer",
            "format": "int64"
          },
          "doctorName": {
            "type": "string"
          },
          "patientId": {
            "type": "integer",
            "format": "int64"
          },
          "patientName": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "DoctorUpdateDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "minLength": 1
          },
          "lastName": {
            "type": "string",
            "minLength": 1
          },
          "specialization": {
            "type": "string",
            "minLength": 1
          },
          "departmentId": {
            "type": "integer",
            "format": "int64"
          },
          "appointmentDuration": {
            "type": "string"
          }
        },
        "required": [
          "appointmentDuration",
          "departmentId",
          "firstName",
          "lastName",
          "specialization"
        ]
      },
      "DoctorDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "specialization": {
            "type": "string"
          },
          "specializationLabel": {
            "type": "string"
          },
          "departmentId": {
            "type": "integer",
            "format": "int64"
          },
          "departmentName": {
            "type": "string"
          },
          "appointmentDuration": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "AppointmentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "appointmentTime": {
            "type": "string",
            "format": "date-time"
          },
          "reason": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "SCHEDULED",
              "COMPLETED",
              "CANCELLED",
              "NO_SHOW"
            ]
          },
          "doctor": {
            "$ref": "#/components/schemas/DoctorDto"
          },
          "patient": {
            "$ref": "#/components/schemas/PatientDto"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "medicalRecord": {
            "$ref": "#/components/schemas/MedicalRecordDto"
          }
        }
      },
...
      "DoctorRegistrationDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "minLength": 1
          },
          "lastName": {
            "type": "string",
            "minLength": 1
          },
          "email": {
            "type": "string",
            "format": "email",
            "minLength": 1
          },
          "password": {
            "type": "string",
            "minLength": 1
          },
          "specialization": {
            "type": "string",
            "minLength": 1
          },
          "departmentId": {
            "type": "integer",
            "format": "int64"
          },
          "appointmentDuration": {
            "type": "string"
          }
        },
        "required": [
          "appointmentDuration",
          "departmentId",
          "email",
          "firstName",
          "lastName",
          "password",
          "specialization"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}